+++ release_notes.md
@@ -7,3 +7,4 @@
 - Increased maximum HTTP request content size to 210000000 Bytes (~200MB)
 - Update Node.js Worker Version to [3.8.1](https://github.com/Azure/azure-functions-nodejs-worker/releases/tag/v3.8.1)
 - Update PowerShell 7.4 Worker Version to [4.0.2930](https://github.com/Azure/azure-functions-powershell-worker/releases/tag/v4.0.2930)
+- Sending command line arguments to language workers with `functions-` prefix to prevent conflicts (#9514)